+++ components/kafka/src/triggers/kafkaConsumer.ts
@@ -1,7 +1,9 @@
 import { trigger, util } from "@prismatic-io/spectral";
 import { createClient } from "../client";
+import { kafkaConsumerExamplePayload } from "../examplePayloads";
 import { kafkaConsumerInputs } from "../inputs";
-import type { KafkaMessage } from "../types/consumer";
+import type { DeserializedValue, KafkaMessage } from "../types/consumer";
+import { createSchemaRegistryClient, deserializeBuffer } from "../utils";
 export const kafkaConsumer = trigger({
   display: {
     label: "Kafka Consumer",
@@ -19,6 +21,7 @@ export const kafkaConsumer = trigger({
       autoCommit,
       sessionTimeout,
       heartbeatInterval,
+      deserializeKeys,
     } = params;
     const kafka = createClient(
       {
@@ -28,6 +31,11 @@ export const kafkaConsumer = trigger({
       },
       context.debug.enabled,
     );
+    const avroEnabled = util.types.toBool(connection.fields.avroEnabled);
+    const registry =
+      avroEnabled && connection.fields.schemaRegistryUrl
+        ? createSchemaRegistryClient(connection)
+        : undefined;
     const consumer = kafka.consumer({
       groupId: consumerGroupId,
       sessionTimeout,
@@ -61,12 +69,32 @@ export const kafkaConsumer = trigger({
               resolveOnce();
               return;
             }
+            let key: DeserializedValue | null = null;
+            let value: DeserializedValue | null = null;
+            if (registry && message.value) {
+              value = await deserializeBuffer(
+                registry,
+                message.value,
+                context.logger,
+              );
+            } else if (message.value) {
+              value = util.types.toString(message.value);
+            }
+            if (registry && deserializeKeys && message.key) {
+              key = await deserializeBuffer(
+                registry,
+                message.key,
+                context.logger,
+              );
+            } else if (message.key) {
+              key = util.types.toString(message.key);
+            }
             messages.push({
               topic,
               partition,
               offset: message.offset,
-              key: message.key ? util.types.toString(message.key) : null,
-              value: message.value ? util.types.toString(message.value) : null,
+              key,
+              value,
               timestamp: message.timestamp,
               headers: message.headers,
             });
@@ -102,4 +130,5 @@ export const kafkaConsumer = trigger({
   inputs: kafkaConsumerInputs,
   scheduleSupport: "required",
   synchronousResponseSupport: "invalid",
+  examplePayload: kafkaConsumerExamplePayload,
 });