+++ components/kafka/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./cleanInput";
 export * from "./misc";
+export * from "./schemaRegistry";